First attempt at correcting x/y swap.
Tablets, lie the TF 300T, have x as the narrow side and are wider than they are tall.  Natraly landscape.  Forcing portate mode should solve this problem for everyone AND be simple.  A more elegant solution is needed, tablet users will have to tilt heds to use the address bar.